# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "MyMechanicShop API"
  description: "API for a mechanic shop with endpoints for customers, mechanic staff, inventory, and service tickets"
  version: '1.0.0'
host: "127.0.0.1:5000"
schemes: 
  - 'http'
consumes:
  - 'application/json'
produces:
  - 'application/json'
securityDefinitions:
  bearerAuthCustomer:
    type: apiKey
    name: Authorization
    in: header
  bearerAuthMechanic:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Customer login endpoint (generates auth token)"
      description: "Validates customer login credentials. Generates and returns a customer specific authentication token that can be used to access restricted endpoints."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login Successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              status: "Success"
              message: "Successfully logged in"
              auth_token: "dfg87EGLsd..."
  /customers:
    post:
      tags:
        - Customers
      summary: "Endpoint to create or register a new customer"
      description: "Create a new customer in the mechanic shop database. Email must be unique. Limit of 3 new customers created per hour."
      parameters:
        - in: "body"
          name: "body"
          description: "Information to Create Customer"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Successfully created customer account"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@email.com"
              phone: "(123)456-7890"
              password: "password123"

    get:
      tags:
        - Customers
      summary: "Returns all customers"
      description: "Endpoint to retrieve all customers from the database (login not required, returns all customer information except the passwords). Data gets cached for 60 seconds. Query parameters for page and per_page can be included in the route to retrieve the data paginated."
      parameters:
        - in: "query"
          name: "page"
          description: "Page number to display of paginated customers"
          required: false
          schema: 
            type: "integer"
        - in: "query"
          name: "per_page"
          description: "Number of customers per page for pagination"
          required: false
          schema: 
            type: "integer"
      responses:
        200:
          description: "Retrieved all customers successfully"
          schema:
            $ref: "#/definitions/AllCustomers"
          examples:
            application/json:
              [
                {
                  id: 1,
                  name: "John Doe",
                  email: "john@email.com",
                  phone: "(123)456-7890"
                },
                {
                  id: 2,
                  name: "Jane Pot",
                  email: "jane@email.com",
                  phone: "(123)456-7891"
                }
              ]

    put:
      tags:
        - Customers
      summary: "Update customer data"
      description: "Endpoint to update the logged-in customer's data. This is a token authenticated route. The customer must be logged in to update their account."
      security:
        - bearerAuthCustomer: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information needed to update customer"
          required: true
          schema: 
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Successfully updated customer"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@email.com"
              phone: "(123)456-7890"
              password: "password123"

    delete:
      tags:
        - Customers
      summary: "Delete customer"
      description: "Endpoint to delete customer. This is a token authenticated route. Customer must be logged in to delete their account. Deleting the account will also delete all service tickets associated with the customer."
      security:
        - bearerAuthCustomer: []
      responses:
        200:
          description: "Successfully deleted customer"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer successfully deleted"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Returns specific customer based on ID"
      description: "Endpoint to retrieve a customer based on the customer's ID (login not required, returns the customer's information except their password)"
      parameters:
        - in: "path"
          name: "customer_id"
          description: "Numeric ID to identify which customer to get"
          required: true
          schema:
            type: "integer"
      responses:
        200:
          description: "Retrieved customer information successfully"
          schema:
            $ref: "#/definitions/SingleCustomerResponse"
          examples:
            application/json:
              id: 1,
              name: "John Doe"
              email: "john@email.com"
              phone: "(123)456-7890"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Retrieve customer's service tickets"
      description: "Retrieves all the service tickets associated with the customer. This is a token authenticated route. Customer must be logged in to see their tickets."
      security:
        - bearerAuthCustomer: []
      responses:
        200:
          description: "Successfully retrieved customer service tickets"
          schema:
            $ref: "#/definitions/AllCustomerTickets"
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "VIN": "1234567",
                  "service_date": "2025-07-22",
                  "service_desc": "Oil change",
                  "customer": {
                    "email": "john@email.com",
                    "name": "John Doe",
                    "phone": "(123)456-7890"
                  },
                  "mechanics": [
                    {
                      "email": "mech@email.com",
                      "name": "Ron",
                      "phone": "(123)456-7892"
                    }
                  ],
                  "items": [
                    {
                      "item": {
                        "name": "item 1",
                        "price": 1.00
                      },
                      "quantity": 1
                    },
                    {
                      "item": {
                        "name": "item 2",
                        "price": 2.00
                      },
                      "quantity": 2
                    }
                  ],
                }
              ]
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: "Mechanic login endpoint (generates auth token)"
      description: "Validates Mechanic login credentials. Generates and returns a Mechanic specific authentication token that can be used to access restricted endpoints."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login Successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              status: "Success"
              message: "Successfully logged in"
              auth_token: "dfg87EGLsd..."
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Endpoint to create or register a new Mechanic"
      description: "Create a new mechanic employee for your mechanic shop. Email must be unique between mechanics. Limit of 10 new mechanics can be created per hour."
      parameters:
        - in: "body"
          name: "body"
          description: "Information to Create Mechanic"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Successfully created mechanic account"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@email.com"
              phone: "(123)456-7890"
              salary: 100000
              password: "password123"

    get:
      tags:
        - Mechanics
      summary: "Returns all mechanics"
      description: "Endpoint to retrieve all mechanics from the database (login not required, returns all mechanic information except the passwords and salaries). Data gets cached for 60 seconds."
      responses:
        200:
          description: "Retrieved all mechanics successfully"
          schema:
            $ref: "#/definitions/AllMechanics"
          examples:
            application/json:
              [
                {
                  id: 1,
                  name: "John Doe",
                  email: "john@email.com",
                  phone: "(123)456-7890"
                },
                {
                  id: 2,
                  name: "Jane Pot",
                  email: "jane@email.com",
                  phone: "(123)456-7891"
                }
              ]

    put:
      tags:
        - Mechanics
      summary: "Update mechanic data"
      description: "Endpoint to update the logged-in mechanic's data. This is a token authenticated route. The mechanic must be logged in to update their account."
      security:
        - bearerAuthMechanic: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information needed to update mechanic"
          required: true
          schema: 
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Successfully updated mechanic"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@email.com"
              phone: "(123)456-7890"
              salary: 100000
              password: "password123"

    delete:
      tags:
        - Mechanics
      summary: "Delete mechanic"
      description: "Endpoint to delete mechanic. This is a token authenticated route. Mechanic must be logged in to delete their account."
      security:
        - bearerAuthMechanic: []
      responses:
        200:
          description: "Successfully deleted mechanic"
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          examples:
            application/json:
              message: "Mechanic successfully deleted"
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      status:
        type: "string"
      message:
        type: "string"
      auth_token:
        type: string
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
  SingleCustomerResponse:
    type: "object"
    properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message: 
        type: "string"
  AllCustomerTickets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        VIN:
          type: "string"
        service_date:
          type: "string"
          format: date
        service_desc:
          type: "string"
        customer:
          type: "object"
          properties:
            name:
              type: "string"
            email:
              type: "string"
            phone:
              type: "string"
        mechanics:
          type: "array"
          items: 
            type: "object"
            properties:
              name:
                type: "string"
              email:
                type: "string"
              phone:
                type: "string"
        items:
          type: "array"
          items:
            type: "object"
            properties:
              item:
                type: "object"
                properties:
                  name:
                    type: "string"
                  price:
                    type: "number"
                    format: "float"
              quantity:
                type: "integer"
  CreateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - salary
      - password
  CreateMechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
  AllMechanics:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
  UpdateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - salary
      - password
  UpdateMechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
  DeleteMechanicResponse:
    type: "object"
    properties:
      message: 
        type: "string"
...